Changes for versions 1.15.4 trough 1.21.0.
Small changes to ESL_Buckaroo to process transaction data the same way, regardless of how it comes to us.
Video support for ESL_Twitter
Status ("tweet") lookup support for ESL_Twitter (required to get embedded entities when working with the search API)
Small improvements to ESL_Images
Addition of ESL_SshTunnel, which allows code to be executed on remote servers as if it is being executed on the local server.